fsck should fix ErrSubvenanceOfProvenance (#5782)
